requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
tqdm>=4.65.0
wandb>=0.15.0
hydra-core>=1.3.0
omegaconf>=2.3.0
pytest>=7.4.0
pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pandas>=2.0.0 